Fixing Compiler Warnings (#2328)
* removed a number of no longer required suppress warning annotations

* fixed some compiler warnings

* fixed some compiler warnings

* fixed some compiler warnings

* fixed some compiler warnings

* fixed some compiler warnings

* fixed some compiler warnings

* minor fix

+++ CoreAlgorithmPlugins/src/au/gov/asd/tac/constellation/plugins/algorithms/sna/similarity/CosineSimilarityPlugin.java
@@ -214,7 +214,7 @@ public class CosineSimilarityPlugin extends SimpleEditPlugin {
         }
 
         private float calculateMagnitude() {
-            float mag = 0;
+            double mag = 0;
             for (final int neighbourWeight : neighbourWeightsMap.values()) {
                 mag += Math.pow(neighbourWeight, 2);
             }
